Broke new encounter form up into smaller chunks
Broke up the new encounter form into a controller class, made it self
contained in the form to try and be consistent with the other forms in
the codebase.

Broke up the actual twig files into lots of sub-modules so pieces can be
overridden by modules.

Next step is to build out an event listener that will allow
customization of the data that gets injected before rendering the form.
Haven't decided if we have a generic TwigRenderEvent for all kinds of
twig data, or have one specific to just the new patient encounter form.
