+++ interface/forms/newpatient/templates/newpatient/partials/common/_head-before.html.twig
@@ -1,0 +1,1 @@
+{# Module Theme writers can override this file to add their own fields #}
