+++ interface/forms/newpatient/templates/newpatient/partials/common/fields/_facility-billing.html.twig
@@ -1,0 +1,16 @@
+{# Billing Facility #}
+{% if not globals.hide_billing_widget %}
+    <div class="col-sm">
+        <div class="form-group">
+            <label for="billing_facility" class="text-right">{{ "Billing Facility"|xlt }}:</label>
+            <select name="billing_facility" id="billing_facility" class="form-control">
+                <option value="">-- {{ "Select Facility"|xlt }} --</option>
+                {% for facility in billingFacilities %}
+                    <option value="{{ facility.id|attr }}" {{ facility.selected ? 'selected' : '' }}>
+                        {{ facility.name|text }}
+                    </option>
+                {% endfor %}
+            </select>
+        </div>
+    </div>
+{% endif %}
