+++ interface/forms/newpatient/templates/newpatient/partials/common/fields/_provider.html.twig
@@ -1,0 +1,14 @@
+{# Provider Selection #}
+<div class="col-sm">
+    <div class="form-group">
+        <label for="provider_id" class="text-right">{{ "Encounter Provider"|xlt }}:</label>
+        {# Note provider_id change will update facility info in javascript #}
+        <select name="provider_id" id="provider_id" class="form-control">
+            {% for provider in providers %}
+                <option value="{{ provider.id|attr }}" {{ provider.selected ? 'selected' : '' }}>
+                    {{ provider.name|text }}
+                </option>
+            {% endfor %}
+        </select>
+    </div>
+</div>
