+++ interface/forms/newpatient/templates/newpatient/partials/common/fields/_reasonForVisit.html.twig
@@ -1,0 +1,8 @@
+<div class="col-sm">
+    <fieldset>
+        <legend>{{ "Reason for Visit"|xlt }}</legend>
+        <div class="form-row mx-3 h-100">
+            <textarea name="reason" id="reason" class="form-control" cols="80" rows="4">{{ encounter.reason|default("")|text }}</textarea>
+        </div>
+    </fieldset>
+</div>
