+++ interface/forms/newpatient/templates/newpatient/partials/common/fields/_referral.html.twig
@@ -1,0 +1,7 @@
+{# Referral Source #}
+<div class="col-sm {{ globals.gbl_visit_referral_source == 1 ? '' : 'd-none' }}">
+    <div class="form-group">
+        <label for="form_referral_source" class="text-right">{{ 'Referral Source'|xlt  }}:</label>
+        {{ selectList('form_referral_source', 'refsource', defaultReferralSource, '', {empty_name: ''}) }}
+    </div>
+</div>
