+++ interface/forms/newpatient/templates/report.html.twig
@@ -1,1 +1,1 @@
-{% include "newpatient/report.html.twig" %}
+{% include "templates/newpatient/report.html.twig" %}
